Extract run subcommand to cmd_run.go
Move run command and related logic to separate file. main.go is now
reduced to 46 lines containing only the root command and file shorthand.
